FEAT: splashscreen

+++ app/src/main/java/id/my/faruq/coffegrader/data/local/entity/ScanDefectEntity.kt
@@ -14,5 +14,5 @@ data class ScanDefectEntity(
 
     val defectName: String,
     val defectCount: Int,
-    val defectValue: Float
+    val defectValue: Double
 )
